requirements.txt
tensorflow-gpu~=2.10.1
matplotlib~=3.6.3
numpy~=1.24.3
Flask~=2.2.3
Pillow~=9.4.0
Werkzeug~=2.2.2